created postLaptops

+++ src/main/java/com/shop/demo/view/entity/PhoneView.java
@@ -16,5 +16,5 @@ public class PhoneView {
     private long id;
     private String name;
     private String color;
-    private float price;
+    private Integer price;
 }
